Relocate SourceProvider-class
