+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/GeneratedEnumProvider.java
@@ -16,14 +16,13 @@
 
 package io.github.chrimle.o2jrm.tests;
 
+import io.github.chrimle.o2jrm.test.meta.filters.GeneratedSourceProvider;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.ArgumentsProvider;
 import org.junit.jupiter.params.support.ParameterDeclarations;
 
-public final class GeneratedEnumProvider extends GeneratedSourceProvider
-    implements ArgumentsProvider {
+public final class GeneratedEnumProvider extends GeneratedSourceProvider {
 
   public static final String GENERATED_ENUM_TESTS_METHOD_SOURCE_CLASS =
       "io.github.chrimle.o2jrm.tests.GeneratedEnumImplTests";
